pyproject.toml
[project]
name = "subo-agent"
version = "0.1.0"
description = "Multi-Agent LangGraph system for Light Airlines"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "chainlit>=1.1.30",
    "faker>=25.0.0",
    "langchain>=0.2.10",
    "langchain-community>=0.2.7",
    "langchain-google-genai>=1.0.7",
    "langchain-huggingface>=0.0.3",
    "langchain-text-splitters>=0.2.2",
    "langgraph>=0.1.8",
    "faiss-cpu>=1.8.0",
    "pypdf>=4.2.0",
    "sentence-transformers>=3.0.0",
    "python-dotenv>=1.0.1",
    "sqlalchemy>=2.0.31",
    "langgraph-cli[inmem]>=0.4.12",
    "langchain-ollama>=1.0.1",
]
